DRT-75 - implemented StatusChartWidths and WidthHelpers classes; - updated SetAuthorWorkSummaryWidths method parameters; - implemented new helper methods in Summary.cs for StatusChart; - implemented GetTotalOriginalEstimate method in TimesheetService; - updated summary partial view with status chart;

+++ JiraReporter/AuthorsProcessing.cs
@@ -306,7 +306,7 @@ namespace JiraReporter
             return max;
         }
 
-        public static void SetAuthorWorkSummaryWidths(Author author, int maxWidth, double maxValue)
+        public static void SetAuthorWorkSummaryWidths(Author author, int maxWidth, int maxValue)
         {
             author.SprintChartPixelWidth = MathHelpers.RuleOfThree(maxWidth, maxValue, (author.SprintWorkedPerDay / 3600));
             author.MonthChartPixelWidth = MathHelpers.RuleOfThree(maxWidth, maxValue, (author.MonthWorkedPerDay / 3600));
